ui-flask/requirements.txt
flask
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1
eventlet==0.35.2
pyserial